Make sandboxing no-op on windows
