+++ internal/runner/service.go
@@ -9,7 +9,6 @@ import (
 	"strings"
 	"time"
 
-	"github.com/Use-Tusk/fence/pkg/fence"
 	"github.com/Use-Tusk/tusk-cli/internal/config"
 	"github.com/Use-Tusk/tusk-cli/internal/log"
 	"github.com/Use-Tusk/tusk-cli/internal/utils"
@@ -68,7 +67,7 @@ func (e *Executor) StartService() error {
 
 	requireSandbox := effectiveSandboxMode == SandboxModeStrict
 	if effectiveSandboxMode != SandboxModeOff && !e.sandboxBypass {
-		if !fence.IsSupported() {
+		if !isSandboxSupported() {
 			if requireSandbox {
 				return fmt.Errorf("strict replay sandbox unavailable: sandbox not supported on this platform")
 			}
@@ -79,76 +78,56 @@ func (e *Executor) StartService() error {
 			if e.getReplaySandboxConfigPath() != "" {
 				sandboxConfigPath = e.getReplaySandboxConfigPath()
 			}
-			fenceCfg, err := createReplayFenceConfig(sandboxConfigPath)
-			if err != nil {
-				return fmt.Errorf("failed to prepare replay sandbox config: %w", err)
-			}
 			if sandboxConfigPath != "" {
 				log.ServiceLog(fmt.Sprintf("🔧 Merged custom Fence config into replay sandbox: %s", utils.ResolveTuskPath(sandboxConfigPath)))
 			}
-			e.fenceManager = fence.NewManager(fenceCfg, e.debug, false)
 
-			// Tell fence how the sandboxed service binds its port. For
-			// docker / docker-compose / podman commands, the daemon binds
-			// the host port outside the sandbox netns, so fence must NOT
-			// set up a reverse bridge (it would collide with the daemon's
-			// bind). For everything else, fence proxies inbound traffic
-			// into the sandbox netns as usual.
-			executionModel := fence.ServiceBindsInSandbox
-			if serviceDelegatesToHostDaemon(cfg.Service.Start.Command) {
-				executionModel = fence.ServiceBindsOnHost
-			}
-			e.fenceManager.SetService(fence.ServiceOptions{
-				ExposedPorts:   []int{cfg.Service.Port},
-				ExecutionModel: executionModel,
-			})
-
-			// Hand any caller-generated host files the sandboxed process
-			// needs to see (e.g. the replay compose env-override YAML)
-			// to fence. Without this, a file created via
-			// os.CreateTemp("", ...) lives under /tmp, which fence
-			// tmpfs-overmounts — invisible to the sandboxed docker client.
-			exposeErr := error(nil)
+			// Build the list of host paths that must be visible inside
+			// the sandbox. The replay compose env-override (when present)
+			// is the main case: it lives under /tmp on the host, which
+			// fence tmpfs-overmounts — without this exposure the sandboxed
+			// docker client can't see the file passed to `-f`.
+			var exposedHostPaths []exposedHostPath
 			if replayOverridePath != "" {
-				if err := e.fenceManager.ExposeHostPath(replayOverridePath, false); err != nil {
-					exposeErr = err
-				}
-			}
-			if exposeErr != nil {
-				if requireSandbox {
-					e.fenceManager = nil
-					return fmt.Errorf("strict replay sandbox unavailable: failed to expose replay override file to sandbox: %w", exposeErr)
-				}
-				log.UserWarn(fmt.Sprintf("⚠️  Sandbox: failed to expose replay override file (%v); proceeding without sandbox", exposeErr))
-				e.fenceManager = nil
+				exposedHostPaths = append(exposedHostPaths, exposedHostPath{
+					Path:     replayOverridePath,
+					Writable: false,
+				})
 			}
 
-			if e.fenceManager != nil {
-				if err := e.fenceManager.Initialize(); err != nil {
+			// For docker / docker-compose / podman commands, the daemon
+			// binds the host port outside the sandbox netns, so the
+			// sandbox must NOT set up a reverse bridge on that port (it
+			// would collide with the daemon's own bind). For everything
+			// else, the sandbox proxies inbound traffic into its netns as
+			// usual.
+			sbx, sbxErr := newReplaySandboxManager(replaySandboxOptions{
+				UserConfigPath:   sandboxConfigPath,
+				Debug:            e.debug,
+				ExposedPort:      cfg.Service.Port,
+				BindsOnHost:      serviceDelegatesToHostDaemon(cfg.Service.Start.Command),
+				ExposedHostPaths: exposedHostPaths,
+			})
+			if sbxErr != nil {
+				if requireSandbox {
+					return fmt.Errorf("strict replay sandbox unavailable: %s", friendlySandboxError(sbxErr))
+				}
+				log.UserWarn(fmt.Sprintf("⚠️  Sandbox unavailable: %s", friendlySandboxError(sbxErr)))
+				log.UserWarn("   Tests will run without network isolation (real connections allowed)\n")
+			} else {
+				wrappedCmd, wrapErr := sbx.WrapCommand(command)
+				if wrapErr != nil {
+					sbx.Cleanup()
 					if requireSandbox {
-						e.fenceManager = nil
-						return fmt.Errorf("strict replay sandbox unavailable: %s", friendlySandboxError(err))
+						return fmt.Errorf("strict replay sandbox unavailable: %s", friendlySandboxError(wrapErr))
 					}
-					log.UserWarn(fmt.Sprintf("⚠️  Sandbox unavailable: %s", friendlySandboxError(err)))
+					log.UserWarn(fmt.Sprintf("⚠️  Sandbox unavailable: %s", friendlySandboxError(wrapErr)))
 					log.UserWarn("   Tests will run without network isolation (real connections allowed)\n")
-					e.fenceManager = nil
 				} else {
-					wrappedCmd, err := e.fenceManager.WrapCommand(command)
-					if err != nil {
-						if requireSandbox {
-							e.fenceManager.Cleanup()
-							e.fenceManager = nil
-							return fmt.Errorf("strict replay sandbox unavailable: %s", friendlySandboxError(err))
-						}
-						log.UserWarn(fmt.Sprintf("⚠️  Sandbox unavailable: %s", friendlySandboxError(err)))
-						log.UserWarn("   Tests will run without network isolation (real connections allowed)\n")
-						e.fenceManager.Cleanup()
-						e.fenceManager = nil
-					} else {
-						command = wrappedCmd
-						e.lastServiceSandboxed = true
-						log.ServiceLog("🔒 Service sandboxed (localhost outbound blocked for replay isolation)")
-					}
+					e.sandbox = sbx
+					command = wrappedCmd
+					e.lastServiceSandboxed = true
+					log.ServiceLog("🔒 Service sandboxed (localhost outbound blocked for replay isolation)")
 				}
 			}
 		}
@@ -223,9 +202,9 @@ func (e *Executor) StartService() error {
 	}
 
 	if err := e.serviceCmd.Start(); err != nil {
-		if e.fenceManager != nil {
-			e.fenceManager.Cleanup()
-			e.fenceManager = nil
+		if e.sandbox != nil {
+			e.sandbox.Cleanup()
+			e.sandbox = nil
 		}
 		return fmt.Errorf("failed to start service: %w", err)
 	}
@@ -250,125 +229,14 @@ func (e *Executor) StartService() error {
 	return nil
 }
 
-// createReplayFenceConfig creates the effective fence config for replay mode.
-// This blocks localhost outbound connections to force the service to use SDK mocks.
-func createReplayFenceConfig(userConfigPath string) (*fence.Config, error) {
-	cfg := baseReplayFenceConfig()
-	if userConfigPath == "" {
-		return cfg, nil
-	}
-
-	resolvedPath := utils.ResolveTuskPath(userConfigPath)
-	userCfg, err := fence.LoadConfigResolved(resolvedPath)
-	if err != nil {
-		return nil, fmt.Errorf("load custom fence config %q: %w", resolvedPath, err)
-	}
-	if userCfg == nil {
-		return nil, fmt.Errorf("custom fence config not found: %s", resolvedPath)
-	}
-	if err := validateReplayFenceConfig(userCfg); err != nil {
-		return nil, err
-	}
-
-	merged := fence.MergeConfigs(cfg, userCfg)
-	applyReplayFenceInvariants(merged)
-	return merged, nil
-}
-
-func baseReplayFenceConfig() *fence.Config {
-	f := false
-	return &fence.Config{
-		Network: fence.NetworkConfig{
-			AllowedDomains: []string{
-				// Allow localhost for the service's own health checks
-				"localhost",
-				"127.0.0.1",
-			},
-			AllowLocalBinding:   true, // Allow service to bind to its port
-			AllowLocalOutbound:  &f,   // Block outbound to localhost (Postgres, Redis, etc.)
-			AllowAllUnixSockets: true, // Allow SDK to connect to mock server via Unix socket
-		},
-		Filesystem: fence.FilesystemConfig{
-			AllowWrite: getAllowedWriteDirs(),
-		},
-	}
-}
-
-func validateReplayFenceConfig(cfg *fence.Config) error {
-	if cfg == nil {
-		return nil
-	}
-
-	requiredDomains := []string{"localhost", "127.0.0.1"}
-	for _, deniedDomain := range cfg.Network.DeniedDomains {
-		for _, requiredDomain := range requiredDomains {
-			if strings.EqualFold(deniedDomain, requiredDomain) {
-				return fmt.Errorf("custom replay fence config cannot deny %q because replay health checks require it", requiredDomain)
-			}
-		}
-	}
-
-	return nil
-}
-
-func applyReplayFenceInvariants(cfg *fence.Config) {
-	if cfg == nil {
-		return
-	}
-
-	f := false
-	cfg.Network.AllowedDomains = mergeUniqueStrings(
-		cfg.Network.AllowedDomains,
-		[]string{"localhost", "127.0.0.1"},
-	)
-	cfg.Network.AllowLocalBinding = true
-	cfg.Network.AllowLocalOutbound = &f
-	cfg.Network.AllowAllUnixSockets = true
-	cfg.Filesystem.AllowWrite = mergeUniqueStrings(cfg.Filesystem.AllowWrite, getAllowedWriteDirs())
-}
-
-func mergeUniqueStrings(existing, required []string) []string {
-	if len(required) == 0 {
-		return existing
-	}
-
-	seen := make(map[string]struct{}, len(existing)+len(required))
-	merged := make([]string, 0, len(existing)+len(required))
-	for _, value := range existing {
-		if _, ok := seen[value]; ok {
-			continue
-		}
-		seen[value] = struct{}{}
-		merged = append(merged, value)
-	}
-	for _, value := range required {
-		if _, ok := seen[value]; ok {
-			continue
-		}
-		seen[value] = struct{}{}
-		merged = append(merged, value)
-	}
-	return merged
-}
-
-// getAllowedWriteDirs returns the default writable paths for replay mode.
-// We allow broad local writes by default. Note that Fence still enforces
-// mandatory dangerous-path protections (see
-// https://github.com/Use-Tusk/fence/blob/main/internal/sandbox/dangerous.go).
-func getAllowedWriteDirs() []string {
-	return []string{
-		"/",
-	}
-}
-
 func (e *Executor) StopService() error {
 	cfg, _ := config.Get()
 
 	defer func() {
 		e.cleanupLogFiles()
-		if e.fenceManager != nil {
-			e.fenceManager.Cleanup()
-			e.fenceManager = nil
+		if e.sandbox != nil {
+			e.sandbox.Cleanup()
+			e.sandbox = nil
 		}
 		// Clean up V8 coverage temp directory
 		if e.coverageTempDir != "" {
